Add files via upload
The above dataset called "train.csv" is used to do predictive analysis and forecast the occurrence/onset of diabetes for women. It consists of data collected with a solution i.e. it has the dependent variable that is already given. So I have just checked out the accuracy and can check out predictions based on other inputs of this type.